Checking local files after permission granted

+++ app/src/main/java/de/welthungerhilfe/cgm/scanner/activities/MainActivity.java
@@ -498,6 +498,9 @@ public class MainActivity extends BaseActivity implements RecyclerDataAdapter.On
         } else {
             File[] children = target.listFiles();
 
+            if (children == null)
+                return;
+
             if (children.length == 0) {
                 target.delete();
             } else {
